+++ tests/testthat/test-huxtable.R
@@ -29,11 +29,11 @@ test_that("markdown caption and notes", {
 
 test_that("save to file", {
 
-  options(modelsummary_html = 'huxtable')
-  options(modelsummary_rtf = 'huxtable')
-  options(modelsummary_latex = 'huxtable')
-  options(modelsummary_word = 'huxtable')
-  options(modelsummary_powerpoint = 'huxtable')
+  options(modelsummary_factory_html = 'huxtable')
+  options(modelsummary_factory_rtf = 'huxtable')
+  options(modelsummary_factory_latex = 'huxtable')
+  options(modelsummary_factory_word = 'huxtable')
+  options(modelsummary_factory_powerpoint = 'huxtable')
 
   random <- random_string()
 
@@ -57,10 +57,10 @@ test_that("save to file", {
   expect_error(modelsummary(models, filename), NA)
   unlink(filename)
 
-  options(modelsummary_html = 'kableExtra')
-  options(modelsummary_rtf = 'gt')
-  options(modelsummary_latex = 'kableExtra')
-  options(modelsummary_word = 'flextable')
-  options(modelsummary_powerpoint = 'flextable')
+  options(modelsummary_factory_html = 'kableExtra')
+  options(modelsummary_factory_rtf = 'gt')
+  options(modelsummary_factory_latex = 'kableExtra')
+  options(modelsummary_factory_word = 'flextable')
+  options(modelsummary_factory_powerpoint = 'flextable')
 
 })